Square Grid + Adjacency
Changed over to using squares for the grid.  On the iPad 3, at 2.0m
size for the edge size, I get 60 FPS with the snake moving around, no
display for the sensors, but displaying the contact counts.  If I go to
1.0m, the FPS drops to 43 (isn) with the contact count display off.
With circular sensors, it is about 40 FPS, so the polygons seem a
little faster for processing the sensor data.

I've added adjacency calculation to the generator.  This will be used
when it is actually put into a graph.