Load Balancing, Part I
I realized that the asteroids were taking about 2-3 FPS out of the
overall performance on the iPad 3.  While they are not really big on
math (a couple of normalize calls), they really don't need to be
updated that often.  I am in the process of creating an EntityScheduler
that will update Entities periodically as needed.  Every Entity will
get it's UpdateDisplay() method called (if it needs it), but it will
only get its Update() method called per the priority it has.