Performance Improvements
I found that the labels for the contact layer were still being updated,
even though they were not being drawn.  This degraded performance.

I added a scheduler for the update of the asteroids.  Now each asteroid
is only updated once per second.  Since they are "kinetic", this is not
an issue.  This got back about 4fps when it was down around 18 (before
other fixes).

I turned off debug drawing.  This gave a big boost, as the drawing of
the sensors was taking lots of cycles.

The big boost/breakpoint though is with the size of the sensors.  By
making the sensors smaller (1.0m) and the grid a little larger (2.5m),
I was able to hit 60 fps.  If the size of the sensors goes up or the
separation a little less (~2.25 m), I see definite drops in fps.  This
is the physics engine and the number of contact updates increasing as
the sizes/separations decrease.  The current size seems like a good fit
for basic path finding.

Next Step:  Pathfinding.