Got it working with Templates.
But it is VERY UGLY.  The real nastiness is in deriving from the
GraphSearchAlgorithm class.  The major data elements are in the base
class, the derived class just handles the algorithm, but I need to use
this-> all over the place so the resolution works properly.  It may be
time for a different approach...