Abandoned Templates (for now)
I committed a bit of c++ heresy by abandoning the use of templates for
now and returning to the old standby: vectors of pointers.  By making
the graph the owner of the pointers and using a common base class for
the search algorithms, the code is much cleaner and a lot less complex.