Got Path Finding Around Obstacles Working!!!
Note:  The ship occupies one of the sensors, so it causes that node to
be "disconnected".  I had to modify the algorithm to allow the start
node to be blocked for the search to work.  Also note that the search
path is stale quickly.  I really need to move the search INTO either
the class or a singleton so that the ship can start a search, wait on
the result, then continually redo the search as it travels till it
reaches the final destination.