Added Flood Filling Display
I added code to show which nodes were explored by the search as well as
the path.  This gives a feel for how much "flooding" the algorithm is
doing.  The BFS floods A LOT as expected.  Next plan is to modify the
Moving Entity so it will do the search dynamically and replan when the
N+2 or N+1 node on the path is occupied (old information in the path
basically).