Looking much better.
I modified the navigation routine so that it checks if the next cell to
be moved into is blocked.  If it is, then it replays the path.  There
is an edge case for the diagonal cells being blocked, which means an
asteroid may be heading towards the ship and usually bangs it on the
back as the ship passes (hitting from side).  I can compute the indexes
for the two adjacent cells (either on diagonal or up/down) based on the
direction of the ship.  This will go into the next round.  Nearly
there….