Implemented the A* Search.
It looks REALLY good.  I thought it looked good when I just used the
Distance heuristic.  The number of nodes expanded relative to Dijkstra
was much smaller than BFS.  When I implemented the Distance Squared
heuristic, I was surprised by how good it really got.  It could be that
I am doing something wrong.  Or it could be that this heuristic really
punishes any exploration that strays too far from the goal path so it
really keeps the search on track.  So, at least for this type of graph,
it seems to give a double "bump" of (1) fewer nodes explored and (2) no
square root hit for evaluating the distance.