Removed Raw IndexedPriorityQueue
I removed the IPQ implementation from Bucland's book and replaced it
with an STL equivalent.  While using Bucland's code as a guide for the
definition of the algorithms (ethically) is reasonable, using his exact
implementation of the IPQ is too close to the gray area for my tastes.
Besides, the IPQ I cooked up is more in line with using the STL the way
it should be used.  The new implementation appears to be faster as
well…this may be because the IPQ previously sorted each time a node was
inserted.  Perhaps the STL version only sorts when top() is called.

I also added metrics to display the number of fill nodes for each
search (in the log).  This shows that the distance squared, distance,
and Manhattan distance are expanding approximately 29, 72, and 44 nodes
respectively on average per search.  So even though the d^2 metric is
not strictly correct (mixing g = distance + h = dist*dist), it works
VERY WELL for this implementation and will be used rolling forward.