Cleaned up the adjacent list generation.
While I was doing this, I also experimented with having adjacent info
for nodes at 1 more out…this generated paths that looked a little more
natural, but the problem was they could path points far enough apart
that the nodes between were occupied even though the ends were not.
For now, it is just better to have standard 8-set adjacency.