Added Smoothing of the path.
I added a Bezier Spline smoother to the patching for the first 5 points
generated.  It looks "OK", but I am not sure it is really all that
helpful.  Also, it appears that the MovingEntity can get hung up
searching for a path because a surrounding node might be blocked, so I
only have it replay if the current node is blocked when traveling into
it.