Fixed bug in smoothing of the path.
The collinear path elimination had a bug.  Now it seems to be working
right.  I think I am going to modify the "in the way" detection so that
it uses "whiskers" instead of occupancy.  The check will occur when a
new path point is pulled, so it should not be a heavy cpu burden when
(I hope).  The current scheme is "ok" but will not prevent collisions
between like objects.