Fixed a few bugs.
The Bezier spline now handles the case where there are only two points
(single line).  Found a bug in the smoothing routine where it was not
checking for the size of path (so putting bad points into Bezier
spline).  Updated the IsNodePassable routine to take a position instead
of a node index.  Also modified the passable check so it looked at the
current node and the next node on the path.